removed screenshot and tighten frame init
The original demo.c saveScreenshot() function directly calls
glReadPixels which is not necessarily available in sokol. And
since I don't want to modify the original code taking a
screenshot has been removed.

Additionally the pxRatio in the frame setup is meant to be
represent hidpi support in nanovg which we have a specific
function in sokol for (sapp_dpi_scale()) so I'm using this one
now which should make the sokol_frame() start a little less
confusing.